requirements.txt
fastapi-restful==0.6.0
fastapi==0.112.1
fhir.resources==7.1.0
gitpython==3.1.43
hypercorn==0.17.3
psycopg2==2.9.9
psycopg2_binary==2.9.9
pyjwt[crypto]==2.9.0
requests==2.32.3
requests_futures==1.0.1
sqlalchemy==2.0.32
typing-inspect==0.9.0